+++ tower/src/util/mod.rs
@@ -1,4 +1,4 @@
-//! Various utility types and functions that are generally with Tower.
+//! Various utility types and functions that are generally used with Tower.
 
 mod and_then;
 mod boxed;
@@ -11,6 +11,7 @@ mod map_request;
 mod map_response;
 mod map_result;
 
+mod map_future;
 mod oneshot;
 mod optional;
 mod ready;
@@ -23,6 +24,7 @@ pub use self::{
     either::Either,
     future_service::{future_service, FutureService},
     map_err::{MapErr, MapErrLayer},
+    map_future::{MapFuture, MapFutureLayer},
     map_request::{MapRequest, MapRequestLayer},
     map_response::{MapResponse, MapResponseLayer},
     map_result::{MapResult, MapResultLayer},
@@ -35,6 +37,8 @@ pub use self::{
 
 pub use self::call_all::{CallAll, CallAllUnordered};
 use std::future::Future;
+
+use crate::layer::util::Identity;
 
 pub mod error {
     //! Error types
@@ -97,12 +101,12 @@ pub trait ServiceExt<Request>: tower_service::Service<Request> {
         CallAll::new(self, reqs)
     }
 
-    /// Executes a new future after this service's after this services future resolves. This does
+    /// Executes a new future after this service's future resolves. This does
     /// not alter the behaviour of the [`poll_ready`] method.
     ///
     /// This method can be used to change the [`Response`] type of the service
     /// into a different type. You can use this method to chain along a computation once the
-    /// services response has been resolved.
+    /// service's response has been resolved.
     ///
     /// [`Response`]: crate::Service::Response
     /// [`poll_ready`]: crate::Service::poll_ready
@@ -171,7 +175,7 @@ pub trait ServiceExt<Request>: tower_service::Service<Request> {
     /// This method can be used to change the [`Response`] type of the service
     /// into a different type. It is similar to the [`Result::map`]
     /// method. You can use this method to chain along a computation once the
-    /// services response has been resolved.
+    /// service's response has been resolved.
     ///
     /// [`Response`]: crate::Service::Response
     /// [`poll_ready`]: crate::Service::poll_ready
@@ -234,7 +238,7 @@ pub trait ServiceExt<Request>: tower_service::Service<Request> {
         MapResponse::new(self, f)
     }
 
-    /// Maps this services's error value to a different value. This does not
+    /// Maps this service's error value to a different value. This does not
     /// alter the behaviour of the [`poll_ready`] method.
     ///
     /// This method can be used to change the [`Error`] type of the service
@@ -859,6 +863,107 @@ pub trait ServiceExt<Request>: tower_service::Service<Request> {
     {
         Then::new(self, f)
     }
+
+    /// Composes a function that transforms futures produced by the service.
+    ///
+    /// This takes a function or closure returning a future computed from the future returned by
+    /// the service's [`call`] method, as opposed to the responses produced by the future.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use std::task::{Poll, Context};
+    /// # use tower::{Service, ServiceExt, BoxError};
+    /// #
+    /// # struct DatabaseService;
+    /// # impl DatabaseService {
+    /// #   fn new(address: &str) -> Self {
+    /// #       DatabaseService
+    /// #   }
+    /// # }
+    /// #
+    /// # type Record = ();
+    /// # type DbError = crate::BoxError;
+    /// #
+    /// # impl Service<u32> for DatabaseService {
+    /// #   type Response = Record;
+    /// #   type Error = DbError;
+    /// #   type Future = futures_util::future::Ready<Result<Record, DbError>>;
+    /// #
+    /// #   fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+    /// #       Poll::Ready(Ok(()))
+    /// #   }
+    /// #
+    /// #   fn call(&mut self, request: u32) -> Self::Future {
+    /// #       futures_util::future::ready(Ok(()))
+    /// #   }
+    /// # }
+    /// #
+    /// # fn main() {
+    /// use std::time::Duration;
+    /// use tokio::time::timeout;
+    ///
+    /// // A service returning Result<Record, DbError>
+    /// let service = DatabaseService::new("127.0.0.1:8080");
+    /// #    async {
+    ///
+    /// let mut new_service = service.map_future(|future| async move {
+    ///     let res = timeout(Duration::from_secs(1), future).await?;
+    ///     Ok::<_, BoxError>(res)
+    /// });
+    ///
+    /// // Call the new service
+    /// let id = 13;
+    /// let record = new_service
+    ///     .ready_and()
+    ///     .await?
+    ///     .call(id)
+    ///     .await?;
+    /// # Ok::<(), BoxError>(())
+    /// #    };
+    /// # }
+    /// ```
+    ///
+    /// Note that normally you wouldn't implement timeouts like this and instead use [`Timeout`].
+    ///
+    /// [`call`]: crate::Service::call
+    /// [`Timeout`]: crate::timeout::Timeout
+    fn map_future<F, Fut, Response, Error>(self, f: F) -> MapFuture<Self, F>
+    where
+        Self: Sized,
+        F: FnMut(Self::Future) -> Fut,
+        Error: From<Self::Error>,
+        Fut: Future<Output = Result<Response, Error>>,
+    {
+        MapFuture::new(self, f)
+    }
 }
 
 impl<T: ?Sized, Request> ServiceExt<Request> for T where T: tower_service::Service<Request> {}
+
+/// Convert an `Option<Layer>` into a [`Layer`].
+///
+/// ```
+/// # use std::time::Duration;
+/// # use tower::Service;
+/// # use tower::builder::ServiceBuilder;
+/// use tower::util::option_layer;
+/// # use tower::timeout::TimeoutLayer;
+/// # async fn wrap<S>(svc: S) where S: Service<(), Error = &'static str> + 'static + Send, S::Future: Send {
+/// # let timeout = Some(Duration::new(10, 0));
+/// // Layer to apply a timeout if configured
+/// let maybe_timeout = option_layer(timeout.map(TimeoutLayer::new));
+///  ServiceBuilder::new()
+///      .layer(maybe_timeout)
+///      .service(svc);
+/// # }
+/// ```
+///
+/// [`Layer`]: crate::layer::Layer
+pub fn option_layer<L>(layer: Option<L>) -> Either<L, Identity> {
+    if let Some(layer) = layer {
+        Either::A(layer)
+    } else {
+        Either::B(Identity::new())
+    }
+}