+++ tower/src/util/service_fn.rs
@@ -1,3 +1,4 @@
+use std::fmt;
 use std::future::Future;
 use std::task::{Context, Poll};
 use tower_service::Service;
@@ -8,9 +9,17 @@ pub fn service_fn<T>(f: T) -> ServiceFn<T> {
 }
 
 /// A [`Service`] implemented by a closure.
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone)]
 pub struct ServiceFn<T> {
     f: T,
+}
+
+impl<T> fmt::Debug for ServiceFn<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("ServiceFn")
+            .field("f", &format_args!("{}", std::any::type_name::<T>()))
+            .finish()
+    }
 }
 
 impl<T, F, Request, R, E> Service<Request> for ServiceFn<T>