+++ tower/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - **util**: Add `ServiceExt::map_future`. ([#542])
+- **util**: Make combinators implement `Debug` in more cases
 - `tracing` is now only pulled in for the features that need it.
 
 [#542]: https://github.com/tower-rs/tower/pull/542