+++ tower/CHANGELOG.md
@@ -9,8 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **util**: Add `ServiceExt::map_future`. ([#542])
 - **util**: Make combinators implement `Debug` in more cases
 - `tracing` is now only pulled in for the features that need it.
+- **util**: Add `option_layer` to convert an `Option<Layer>` into a `Layer`. ([#555])
+- **builder**: Add `ServiceBuilder::option_layer` to optionally add a layer. ([#555])
 
 [#542]: https://github.com/tower-rs/tower/pull/542
+[#555]: https://github.com/tower-rs/tower/pull/555
 
 # 0.4.4 (January 20, 2021)