+++ tower/CHANGELOG.md
@@ -6,10 +6,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-### Added
-- **util**: Implement `Layer` for `Either<A, B>`.
+- **util**: Add `ServiceExt::map_future`. ([#542])
+- **util**: Make combinators implement `Debug` in more cases
+- `tracing` is now only pulled in for the features that need it.
+- **util**: Add `option_layer` to convert an `Option<Layer>` into a `Layer`. ([#555])
+- **builder**: Add `ServiceBuilder::option_layer` to optionally add a layer. ([#555])
+- **steer**: `Steer` now implements `Clone
 - **make**: Added `Shared` which lets you implement `MakeService` by cloning a
   service.
+
+[#542]: https://github.com/tower-rs/tower/pull/542
+[#555]: https://github.com/tower-rs/tower/pull/555
+
+# 0.4.4 (January 20, 2021)
+
+### Added
+
+- **util**: Implement `Layer` for `Either<A, B>`. ([#531])
+- **util**: Implement `Clone` for `FilterLayer`. ([#535])
+- **timeout**: Implement `Clone` for `TimeoutLayer`. ([#535])
+- **limit**: Implement `Clone` for `RateLimitLayer`. ([#535])
+
+### Fixed
+
+- Added "full" feature which turns on all other features. ([#532])
+- **spawn-ready**: Avoid oneshot allocations. ([#538])
+
+[#531]: https://github.com/tower-rs/tower/pull/531
+[#532]: https://github.com/tower-rs/tower/pull/532
+[#535]: https://github.com/tower-rs/tower/pull/535
+[#538]: https://github.com/tower-rs/tower/pull/538
+>>>>>>> master
 
 # 0.4.3 (January 13, 2021)